+++ src/types/stardust/block/output/unlock_condition/storage_deposit_return.rs
@@ -7,7 +7,7 @@ use bee_block_stardust::output::unlock_condition as bee;
 use serde::{Deserialize, Serialize};
 
 use super::OutputAmount;
-use crate::types::stardust::block::Address;
+use crate::types::{context::TryFromWithContext, stardust::block::Address};
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct StorageDepositReturnUnlockCondition {
@@ -24,11 +24,14 @@ impl<T: Borrow<bee::StorageDepositReturnUnlockCondition>> From<T> for StorageDep
     }
 }
 
-impl TryFrom<StorageDepositReturnUnlockCondition> for bee::StorageDepositReturnUnlockCondition {
+impl TryFromWithContext<StorageDepositReturnUnlockCondition> for bee::StorageDepositReturnUnlockCondition {
     type Error = bee_block_stardust::Error;
 
-    fn try_from(value: StorageDepositReturnUnlockCondition) -> Result<Self, Self::Error> {
-        bee::StorageDepositReturnUnlockCondition::new(value.return_address.into(), value.amount.0)
+    fn try_from_with_context(
+        ctx: &bee_block_stardust::protocol::ProtocolParameters,
+        value: StorageDepositReturnUnlockCondition,
+    ) -> Result<Self, Self::Error> {
+        bee::StorageDepositReturnUnlockCondition::new(value.return_address.into(), value.amount.0, ctx.token_supply())
     }
 }
 
@@ -38,10 +41,10 @@ mod rand {
 
     impl StorageDepositReturnUnlockCondition {
         /// Generates a random [`StorageDepositReturnUnlockCondition`].
-        pub fn rand() -> Self {
+        pub fn rand(ctx: &bee_block_stardust::protocol::ProtocolParameters) -> Self {
             Self {
                 return_address: Address::rand_ed25519(),
-                amount: OutputAmount::rand(),
+                amount: OutputAmount::rand(ctx),
             }
         }
     }