+++ src/types/stardust/block/payload/milestone/mod.rs
@@ -10,6 +10,7 @@ use serde::{Deserialize, Serialize};
 
 pub use self::milestone_id::MilestoneId;
 use crate::types::{
+    context::{TryFromWithContext, TryIntoWithContext},
     stardust::block::{payload::treasury_transaction::TreasuryTransactionPayload, Address, BlockId, Signature},
     tangle::MilestoneIndex,
     util::bytify,
@@ -30,12 +31,15 @@ impl<T: Borrow<bee::MilestonePayload>> From<T> for MilestonePayload {
     }
 }
 
-impl TryFrom<MilestonePayload> for bee::MilestonePayload {
+impl TryFromWithContext<MilestonePayload> for bee::MilestonePayload {
     type Error = bee_block_stardust::Error;
 
-    fn try_from(value: MilestonePayload) -> Result<Self, Self::Error> {
+    fn try_from_with_context(
+        ctx: &bee_block_stardust::protocol::ProtocolParameters,
+        value: MilestonePayload,
+    ) -> Result<Self, Self::Error> {
         bee::MilestonePayload::new(
-            value.essence.try_into()?,
+            value.essence.try_into_with_context(ctx)?,
             Vec::from(value.signatures)
                 .into_iter()
                 .map(Into::into)
@@ -44,11 +48,14 @@ impl TryFrom<MilestonePayload> for bee::MilestonePayload {
     }
 }
 
-impl TryFrom<MilestonePayload> for bee::dto::MilestonePayloadDto {
+impl TryFromWithContext<MilestonePayload> for bee::dto::MilestonePayloadDto {
     type Error = bee_block_stardust::Error;
 
-    fn try_from(value: MilestonePayload) -> Result<Self, Self::Error> {
-        let stardust = bee::MilestonePayload::try_from(value)?;
+    fn try_from_with_context(
+        ctx: &bee_block_stardust::protocol::ProtocolParameters,
+        value: MilestonePayload,
+    ) -> Result<Self, Self::Error> {
+        let stardust = bee::MilestonePayload::try_from_with_context(ctx, value)?;
         Ok(Self::from(&stardust))
     }
 }
@@ -88,13 +95,17 @@ impl<T: Borrow<bee::MilestoneEssence>> From<T> for MilestoneEssence {
     }
 }
 
-impl TryFrom<MilestoneEssence> for bee::MilestoneEssence {
+impl TryFromWithContext<MilestoneEssence> for bee::MilestoneEssence {
     type Error = bee_block_stardust::Error;
 
-    fn try_from(value: MilestoneEssence) -> Result<Self, Self::Error> {
+    fn try_from_with_context(
+        ctx: &bee_block_stardust::protocol::ProtocolParameters,
+        value: MilestoneEssence,
+    ) -> Result<Self, Self::Error> {
         bee::MilestoneEssence::new(
             value.index.into(),
             value.timestamp,
+            ctx.protocol_version(),
             value.previous_milestone_id.into(),
             bee_block_stardust::parent::Parents::new(
                 Vec::from(value.parents).into_iter().map(Into::into).collect::<Vec<_>>(),
@@ -105,7 +116,7 @@ impl TryFrom<MilestoneEssence> for bee::MilestoneEssence {
             bee_block_stardust::payload::MilestoneOptions::new(
                 Vec::from(value.options)
                     .into_iter()
-                    .map(TryInto::try_into)
+                    .map(|x| x.try_into_with_context(ctx))
                     .collect::<Result<Vec<_>, _>>()?,
             )?,
         )
@@ -146,10 +157,13 @@ impl<T: Borrow<bee::MilestoneOption>> From<T> for MilestoneOption {
     }
 }
 
-impl TryFrom<MilestoneOption> for bee::MilestoneOption {
+impl TryFromWithContext<MilestoneOption> for bee::MilestoneOption {
     type Error = bee_block_stardust::Error;
 
-    fn try_from(value: MilestoneOption) -> Result<Self, Self::Error> {
+    fn try_from_with_context(
+        ctx: &bee_block_stardust::protocol::ProtocolParameters,
+        value: MilestoneOption,
+    ) -> Result<Self, Self::Error> {
         Ok(match value {
             MilestoneOption::Receipt {
                 migrated_at,
@@ -161,9 +175,10 @@ impl TryFrom<MilestoneOption> for bee::MilestoneOption {
                 last,
                 Vec::from(funds)
                     .into_iter()
-                    .map(TryInto::try_into)
+                    .map(|x| x.try_into_with_context(ctx))
                     .collect::<Result<Vec<_>, _>>()?,
-                transaction.try_into()?,
+                transaction.try_into_with_context(ctx)?,
+                ctx.token_supply(),
             )?),
             MilestoneOption::Parameters {
                 target_milestone_index,
@@ -203,14 +218,18 @@ impl<T: Borrow<bee::option::MigratedFundsEntry>> From<T> for MigratedFundsEntry 
     }
 }
 
-impl TryFrom<MigratedFundsEntry> for bee::option::MigratedFundsEntry {
+impl TryFromWithContext<MigratedFundsEntry> for bee::option::MigratedFundsEntry {
     type Error = bee_block_stardust::Error;
 
-    fn try_from(value: MigratedFundsEntry) -> Result<Self, Self::Error> {
+    fn try_from_with_context(
+        ctx: &bee_block_stardust::protocol::ProtocolParameters,
+        value: MigratedFundsEntry,
+    ) -> Result<Self, Self::Error> {
         Self::new(
             bee::option::TailTransactionHash::new(value.tail_transaction_hash)?,
             value.address.into(),
             value.amount,
+            ctx.token_supply(),
         )
     }
 }
@@ -226,14 +245,14 @@ mod rand {
 
     impl MilestonePayload {
         /// Generates a random [`MilestonePayload`].
-        pub fn rand() -> Self {
-            rand_milestone_payload().into()
+        pub fn rand(ctx: &bee_block_stardust::protocol::ProtocolParameters) -> Self {
+            rand_milestone_payload(ctx.protocol_version()).into()
         }
     }
 
     impl MilestoneEssence {
         /// Generates a random [`MilestoneEssence`].
-        pub fn rand() -> Self {
+        pub fn rand(ctx: &bee_block_stardust::protocol::ProtocolParameters) -> Self {
             Self {
                 index: rand_number::<u32>().into(),
                 timestamp: rand_number::<u32>(),
@@ -242,15 +261,15 @@ mod rand {
                 inclusion_merkle_root: *rand_merkle_root(),
                 applied_merkle_root: *rand_merkle_root(),
                 metadata: rand_bytes(32),
-                options: Box::new([MilestoneOption::rand_receipt()]),
+                options: Box::new([MilestoneOption::rand_receipt(ctx)]),
             }
         }
     }
 
     impl MilestoneOption {
         /// Generates a random receipt [`MilestoneOption`].
-        pub fn rand_receipt() -> Self {
-            bee::MilestoneOption::from(rand_receipt_milestone_option()).into()
+        pub fn rand_receipt(ctx: &bee_block_stardust::protocol::ProtocolParameters) -> Self {
+            bee::MilestoneOption::from(rand_receipt_milestone_option(ctx.token_supply())).into()
         }
 
         /// Generates a random parameters [`MilestoneOption`].
@@ -265,8 +284,8 @@ mod rand {
 
     impl MigratedFundsEntry {
         /// Generates a random [`MigratedFundsEntry`].
-        pub fn rand() -> Self {
-            rand_migrated_funds_entry().into()
+        pub fn rand(ctx: &bee_block_stardust::protocol::ProtocolParameters) -> Self {
+            rand_migrated_funds_entry(ctx.token_supply()).into()
         }
     }
 }
@@ -287,8 +306,9 @@ mod test {
 
     #[test]
     fn test_milestone_payload_bson() {
-        let payload = MilestonePayload::rand();
-        bee::MilestonePayload::try_from(payload.clone()).unwrap();
+        let ctx = bee_block_stardust::protocol::protocol_parameters();
+        let payload = MilestonePayload::rand(&ctx);
+        bee::MilestonePayload::try_from_with_context(&ctx, payload.clone()).unwrap();
         let bson = to_bson(&payload).unwrap();
         assert_eq!(payload, from_bson::<MilestonePayload>(bson).unwrap());
     }