+++ src/types/stardust/block/payload/treasury_transaction.rs
@@ -7,6 +7,7 @@ use bee_block_stardust::payload as bee;
 use serde::{Deserialize, Serialize};
 
 use super::milestone::MilestoneId;
+use crate::types::context::TryFromWithContext;
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct TreasuryTransactionPayload {
@@ -24,13 +25,16 @@ impl<T: Borrow<bee::TreasuryTransactionPayload>> From<T> for TreasuryTransaction
     }
 }
 
-impl TryFrom<TreasuryTransactionPayload> for bee::TreasuryTransactionPayload {
+impl TryFromWithContext<TreasuryTransactionPayload> for bee::TreasuryTransactionPayload {
     type Error = bee_block_stardust::Error;
 
-    fn try_from(value: TreasuryTransactionPayload) -> Result<Self, Self::Error> {
+    fn try_from_with_context(
+        ctx: &bee_block_stardust::protocol::ProtocolParameters,
+        value: TreasuryTransactionPayload,
+    ) -> Result<Self, Self::Error> {
         Self::new(
             bee_block_stardust::input::TreasuryInput::new(value.input_milestone_id.into()),
-            bee_block_stardust::output::TreasuryOutput::new(value.output_amount)?,
+            bee_block_stardust::output::TreasuryOutput::new(value.output_amount, ctx.token_supply())?,
         )
     }
 }
@@ -43,8 +47,8 @@ mod rand {
 
     impl TreasuryTransactionPayload {
         /// Generates a random [`TreasuryTransactionPayload`].
-        pub fn rand() -> Self {
-            rand_treasury_transaction_payload().into()
+        pub fn rand(ctx: &bee_block_stardust::protocol::ProtocolParameters) -> Self {
+            rand_treasury_transaction_payload(ctx.token_supply()).into()
         }
     }
 }
@@ -57,8 +61,9 @@ mod test {
 
     #[test]
     fn test_treasury_transaction_payload_bson() {
-        let payload = TreasuryTransactionPayload::rand();
-        bee::TreasuryTransactionPayload::try_from(payload).unwrap();
+        let ctx = bee_block_stardust::protocol::protocol_parameters();
+        let payload = TreasuryTransactionPayload::rand(&ctx);
+        bee::TreasuryTransactionPayload::try_from_with_context(&ctx, payload).unwrap();
         let bson = to_bson(&payload).unwrap();
         assert_eq!(payload, from_bson::<TreasuryTransactionPayload>(bson).unwrap());
     }